Module Tezos_time_measurement_runtime

module Clock : sig ... end

module Default : sig ... end

PLEASE, DO NOT DIRECTLY USE THIS MODULE BY YOURSELF IN YOUR OWN CODE. IT IS ONLY INTENDED TO BE USED THROUGH Tezos_time_measurement_ppx PPX REWRITERS FOR TESTING.

module Measurement : sig ... end

This module defines the model of a measurement as well as its related pretty printers and encoders.

module Publisher : sig ... end

module State : sig ... end

PLEASE, DO NOT DIRECTLY USE THIS MODULE BY YOURSELF IN YOUR OWN CODE. IT IS ONLY INTENDED TO BE USED THROUGH Tezos_time_measurement_ppx PPX REWRITERS FOR TESTING.

module Time_measurement : sig ... end