Module Tezos_time_measurement_runtime.Default

PLEASE, DO NOT DIRECTLY USE THIS MODULE BY YOURSELF IN YOUR OWN CODE. IT IS ONLY INTENDED TO BE USED THROUGH Tezos_time_measurement_ppx PPX REWRITERS FOR TESTING.

AN UNWISE USE COULD SEVERELY IMPACT THE MEMORY USAGE OF YOUR PROGRAM.

This module is the entry point of this library. It is used to perform dependency injection between its modules. It regroups default implementations that must be used at runtime.

module Time_measurement : Time_measurement.S

Default implementation of the Time_measurement.S interface.