Module Tezos_time_measurement_runtime.State

PLEASE, DO NOT DIRECTLY USE THIS MODULE BY YOURSELF IN YOUR OWN CODE. IT IS ONLY INTENDED TO BE USED THROUGH Tezos_time_measurement_ppx PPX REWRITERS FOR TESTING.

AN UNWISE USE COULD SEVERELY IMPACT THE MEMORY USAGE OF YOUR PROGRAM.

module type Elt = sig ... end

Defines the type of an element inside the state.

module type S = sig ... end

Defines an interface to interact with a mutable state that could aggregate values.

module WithRef (E : Elt) : S with type elt := E.t

State.S default implementation relying on Ocaml ref. The state is initialized with the given Elt.init value.