Module type State.S

Defines an interface to interact with a mutable state that could aggregate values.

type elt

The type of elements in the state.

val get_all_and_reset : unit -> elt list

get_all_and_reset () evaluates in all the elements that were aggregated inside the state and then resets the state to its initial value. Elements are returned in the order they were pushed in the state (the resulting list starts by older elements).

val push : elt -> unit

push elt aggregates the given element elt in the state.