Module Tezos_time_measurement_runtime.Time_measurement

A loggable batch of time measurements

PLEASE, DO NOT DIRECTLY USE THIS MODULE BY YOURSELF IN YOUR OWN CODE. IT IS ONLY INTENDED TO BE USED THROUGH Tezos_time_measurement_ppx PPX REWRITERS FOR TESTING.

AN UNWISE USE COULD SEVERELY IMPACT THE MEMORY USAGE OF YOUR PROGRAM.

This module offers several functionalities:

Measuring time of Ocaml expression execution.

Mesuring the current time just before executing an Ocaml expression.

Labelling these mesurements and storing them.

Logging all stored measurements as a batch.

module type S = sig ... end

module Make (S : State.S with type elt := Measurement.t) (P : Publisher.S) (C : Clock.S) : S

A time measurement tool that uses the given clock C to measure the time, stores resulting measurements inside the given state S and publishes stored measurements using the publisher P.